Cast namespace to symbol when parsing event data
The worker is looking for a namespace field in the event 
and uses it if it is available. However, the namespace 
needs to be a symbol and there is no way to currently 
represent that is JSON. So, we have to explicitly parse 
the namespace in the event as a symbol (if there is one).