change the mechanics of field_values_at
This changes field_values_at a bit. it's focused on providing the first n results, and then wrapping the last one as an 'Others' column. This enables users to have graphs that kinda make sense, without overwhelming the number of columns.

it also reduces the max_fields down to 25 by default, as rendering 50 is visually disturbing for anything other than single integer labels :(